+++ test/integration/db/db_impl_test.dart
@@ -49,9 +49,6 @@ import 'dart:async';
 import 'package:test/test.dart';
 
 import 'package:gcloud/db.dart' as db;
-import 'package:appengine/src/appengine_context.dart';
-import 'package:appengine/src/api_impl/raw_datastore_v3_impl.dart';
-import 'package:appengine/src/protobuf_api/rpc/rpc_service_remote_api.dart';
 
 @db.Kind()
 class Person extends db.Model {
@@ -676,14 +673,4 @@ Future waitUntilEntitiesHelper(db.DatastoreDB mdb,
   return Future.forEach(keysByKind.keys.toList(), (Type kind) {
     return waitForKeys(kind, keysByKind[kind]);
   });
-}
-
-void main() {
-  var rpcService = new RPCServiceRemoteApi('127.0.0.1', 4444);
-  var appengineContext = new AppengineContext(
-      'dev', 'test-application', 'test-version', null, null, null);
-  var datastore =
-      new DatastoreV3RpcImpl(rpcService, appengineContext, '<invalid-ticket>');
-
-  runTests(new db.DatastoreDB(datastore), null);
 }